+++ src/aux-records/AuthController.spec.ts
@@ -101,7 +101,6 @@ describe('AuthController', () => {
                         featureList: [],
                         purchasable: true,
                         tier: 'beta',
-                        defaultSubscription: true,
                     },
                 ],
                 webhookSecret: 'webhook',
@@ -2634,6 +2633,56 @@ describe('AuthController', () => {
                         subscriptionId: 'sub_2',
                     });
                 });
+
+                it('should use the default subscription if the user doesnt have one', async () => {
+                    store.subscriptionConfiguration.subscriptions[1].defaultSubscription =
+                        true;
+                    await store.saveUser({
+                        id: 'myid',
+                        email: 'email',
+                        phoneNumber: 'phonenumber',
+                        allSessionRevokeTimeMs: undefined,
+                        currentLoginRequestId: undefined,
+                    });
+
+                    const requestId = 'requestId';
+                    const sessionId = toBase64String('sessionId');
+                    const code = 'code';
+                    const userId = 'myid';
+
+                    const sessionKey = formatV1SessionKey(
+                        userId,
+                        sessionId,
+                        code,
+                        200
+                    );
+
+                    await store.saveSession({
+                        requestId,
+                        sessionId,
+                        secretHash: hashPasswordWithSalt(code, sessionId),
+                        connectionSecret: code,
+                        expireTimeMs: 200,
+                        grantedTimeMs: 100,
+                        previousSessionId: null,
+                        nextSessionId: null,
+                        revokeTimeMs: null,
+                        userId,
+                        ipAddress: '127.0.0.1',
+                    });
+
+                    const result = await controller.validateSessionKey(
+                        sessionKey
+                    );
+
+                    expect(result).toEqual({
+                        success: true,
+                        userId: userId,
+                        sessionId: sessionId,
+                        subscriptionTier: 'beta',
+                        subscriptionId: 'sub_1',
+                    });
+                });
             });
 
             describe('v2 hashes', () => {
@@ -2769,6 +2818,59 @@ describe('AuthController', () => {
                         sessionId: sessionId,
                         subscriptionTier: 'alpha',
                         subscriptionId: 'sub_2',
+                    });
+                });
+
+                it('should use the default subscription if the user doesnt have one', async () => {
+                    store.subscriptionConfiguration.subscriptions[1].defaultSubscription =
+                        true;
+                    await store.saveUser({
+                        id: 'myid',
+                        email: 'email',
+                        phoneNumber: 'phonenumber',
+                        allSessionRevokeTimeMs: undefined,
+                        currentLoginRequestId: undefined,
+                    });
+
+                    const requestId = 'requestId';
+                    const sessionId = toBase64String('sessionId');
+                    const code = 'code';
+                    const userId = 'myid';
+
+                    const sessionKey = formatV1SessionKey(
+                        userId,
+                        sessionId,
+                        code,
+                        200
+                    );
+
+                    await store.saveSession({
+                        requestId,
+                        sessionId,
+                        secretHash: hashHighEntropyPasswordWithSalt(
+                            code,
+                            sessionId
+                        ),
+                        connectionSecret: code,
+                        expireTimeMs: 200,
+                        grantedTimeMs: 100,
+                        previousSessionId: null,
+                        nextSessionId: null,
+                        revokeTimeMs: null,
+                        userId,
+                        ipAddress: '127.0.0.1',
+                    });
+
+                    const result = await controller.validateSessionKey(
+                        sessionKey
+                    );
+
+                    expect(result).toEqual({
+                        success: true,
+                        userId: userId,
+                        sessionId: sessionId,
+                        subscriptionTier: 'beta',
+                        subscriptionId: 'sub_1',
                     });
                 });
             });
@@ -3138,6 +3240,58 @@ describe('AuthController', () => {
                     connectionId: 'connectionId',
                     recordName: 'recordName',
                     inst: 'inst',
+                });
+            });
+
+            it('should give the user the default subscription', async () => {
+                store.subscriptionConfiguration.subscriptions[1].defaultSubscription =
+                    true;
+                const requestId = 'requestId';
+                const sessionId = toBase64String('sessionId');
+                const code = 'code';
+                const connectionSecret = 'connectionSecret';
+                const userId = 'myid';
+
+                await store.saveSession({
+                    requestId,
+                    sessionId,
+                    secretHash: hashHighEntropyPasswordWithSalt(
+                        code,
+                        sessionId
+                    ),
+                    connectionSecret: toBase64String(connectionSecret),
+                    expireTimeMs: 200,
+                    grantedTimeMs: 100,
+                    previousSessionId: null,
+                    nextSessionId: null,
+                    revokeTimeMs: null,
+                    userId,
+                    ipAddress: '127.0.0.1',
+                });
+
+                const connectionKey = formatV1ConnectionKey(
+                    userId,
+                    sessionId,
+                    toBase64String(connectionSecret),
+                    200
+                );
+                const token = generateV1ConnectionToken(
+                    connectionKey,
+                    'connectionId',
+                    'recordName',
+                    'inst'
+                );
+                const result = await controller.validateConnectionToken(token);
+
+                expect(result).toEqual({
+                    success: true,
+                    userId: userId,
+                    sessionId: sessionId,
+                    connectionId: 'connectionId',
+                    recordName: 'recordName',
+                    inst: 'inst',
+                    subscriptionTier: 'beta',
+                    subscriptionId: 'sub_1',
                 });
             });
 
@@ -4589,6 +4743,72 @@ describe('AuthController', () => {
                     publishProjects: true,
                     projectDevelopment: true,
                 },
+            });
+        });
+
+        it('should include use the default subscription if configured', async () => {
+            store.subscriptionConfiguration.subscriptions[1].defaultSubscription =
+                true;
+            await store.saveUser({
+                id: userId,
+                email: 'email',
+                phoneNumber: 'phonenumber',
+                name: 'Test',
+                avatarUrl: 'avatar url',
+                avatarPortraitUrl: 'avatar portrait url',
+                allSessionRevokeTimeMs: undefined,
+                currentLoginRequestId: undefined,
+            });
+
+            const result = await controller.getUserInfo({
+                userId,
+                sessionKey,
+            });
+
+            expect(result).toEqual({
+                success: true,
+                userId: userId,
+                email: 'email',
+                phoneNumber: 'phonenumber',
+                name: 'Test',
+                avatarUrl: 'avatar url',
+                avatarPortraitUrl: 'avatar portrait url',
+                hasActiveSubscription: false,
+                subscriptionTier: 'beta',
+            });
+        });
+
+        it('should include use the default subscription if the user has an inactive subscription', async () => {
+            store.subscriptionConfiguration.subscriptions[1].defaultSubscription =
+                true;
+            await store.saveUser({
+                id: userId,
+                email: 'email',
+                phoneNumber: 'phonenumber',
+                name: 'Test',
+                avatarUrl: 'avatar url',
+                avatarPortraitUrl: 'avatar portrait url',
+                subscriptionStatus: 'canceled',
+                subscriptionId: 'sub_2',
+                allSessionRevokeTimeMs: undefined,
+                currentLoginRequestId: undefined,
+            });
+
+            const result = await controller.getUserInfo({
+                userId,
+                sessionKey,
+            });
+
+            expect(result).toEqual({
+                success: true,
+                userId: userId,
+                email: 'email',
+                phoneNumber: 'phonenumber',
+                name: 'Test',
+                avatarUrl: 'avatar url',
+                avatarPortraitUrl: 'avatar portrait url',
+                hasActiveSubscription: false,
+                subscriptionTier: 'beta',
             });
         });