+++ src/aux-records/RecordsServer.ts
@@ -3945,6 +3945,7 @@ export class RecordsServer {
                     description: s.description,
                     featureList: s.featureList,
                     prices: s.prices,
+                    defaultSubscription: s.defaultSubscription,
                 })
             ),
         });
@@ -4028,6 +4029,7 @@ export class RecordsServer {
                     description: s.description,
                     featureList: s.featureList,
                     prices: s.prices,
+                    defaultSubscription: s.defaultSubscription,
                 })
             ),
         });