+++ src/aux-runtime/runtime/AuxLibrary.ts
@@ -3230,6 +3230,7 @@ export function createDefaultLibrary(context: AuxGlobalContext) {
                 appHooks: { ...hooks, render },
                 listBuiltinTags,
                 requestAuthBot,
+                requestAuthBotInBackground,
 
                 getPublicRecordKey,
                 getSubjectlessPublicRecordKey,
@@ -7846,6 +7847,8 @@ export function createDefaultLibrary(context: AuxGlobalContext) {
      *
      * See [Auth Bot Tags](page:tags#auth-bot-tags) for more information.
      *
+     * See {@link os.requestAuthBotInBackground} for a version of this function that does not show a popup if the user is not signed in.
+     *
      * @example Request an auth bot for the user
      * await os.requestAuthBot();
      * os.toast("Logged in!");
@@ -7857,8 +7860,46 @@ export function createDefaultLibrary(context: AuxGlobalContext) {
      * @docgroup 01-records
      * @docname os.requestAuthBot
      */
-    async function requestAuthBot(): Promise<Bot> {
-        const data = await requestAuthData();
+    function requestAuthBot(): Promise<Bot> {
+        return _requestAuthBot(false);
+    }
+
+    /**
+     * Requests that an "authentication" bot be added to the inst for the current browser tab.
+     * Works similarly to {@link os.requestAuthBot}, except that the request will not show a popup if the user is not signed in.
+     *
+     * Auth bots are useful for discovering general information about the logged in user and are typically associated with a [https://publicos.link](https://publicos.link) user account.
+     *
+     * Returns a promise that resolves with a bot that contains information about the signed in user session.
+     * Resolves with `null` if the user is not already signed in.
+     *
+     * On success, the `authBot` global variable will reference the bot that was returned by the promise.
+     *
+     * See [Auth Bot Tags](page:tags#auth-bot-tags) for more information.
+     *
+     * See {@link os.requestAuthBot} for a version of this function that shows a popup if the user is not signed in.
+     *
+     * @example Request the auth bot in the background.
+     * const authBot = await os.requestAuthBotInBackground();
+     * if (authBot) {
+     *     os.toast("Logged in!");
+     * } else {
+     *     os.toast("Not logged in.");
+     * }
+     *
+     * @dochash actions/records
+     * @doctitle Records Actions
+     * @docsidebar Records
+     * @docdescription Records are a way to store permenent data in CasualOS.
+     * @docgroup 01-records
+     * @docname os.requestAuthBotInBackground
+     */
+    function requestAuthBotInBackground(): Promise<Bot> {
+        return _requestAuthBot(true);
+    }
+
+    async function _requestAuthBot(background: boolean) {
+        const data = await requestAuthData(background);
 
         if (!data) {
             return null;
@@ -7886,9 +7927,9 @@ export function createDefaultLibrary(context: AuxGlobalContext) {
         return bot;
     }
 
-    function requestAuthData(): Promise<AuthData> {
+    function requestAuthData(background: boolean): Promise<AuthData> {
         const task = context.createTask();
-        const event = calcRequestAuthData(task.taskId);
+        const event = calcRequestAuthData(background, task.taskId);
         return addAsyncAction(task, event);
     }